refactor: extract routes and middleware
